requirements.txt
absl-py==0.15.0
addict==2.4.0
albumentations==0.4.6
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
Brlapi==0.6.6
cached-property==1.5.2
cachetools==4.2.4
carla==0.9.5
certifi==2018.1.18
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
clang==5.0
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dill==0.3.4
distro-info===0.18ubuntu0.18.04.1
docker==2.5.1
docker-pycreds==0.2.1
dpkt==1.9.7.2
efficientnet-pytorch==0.6.3
entrypoints==0.4
fail2ban==0.10.2
filelock==3.4.1
flatbuffers==1.12
gast==0.4.0
gdown==4.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.44.0
h5py==3.1.0
httplib2==0.9.2
huggingface-hub==0.4.0
idna==2.6
imageio==2.15.0
imgaug==0.4.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
install==1.3.5
iotop==0.6
ipykernel==5.5.6
ipython==7.16.3
ipython_genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
Jinja2==3.0.3
jiwer==2.3.0
joblib==1.1.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
libpcap==1.11.0b2
librosa==0.9.1
llvmlite==0.36.0
louis==3.5.0
lxml==4.6.2
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mmcv==1.2.5
mmcv-full==1.5.3
multiprocess==0.70.12.2
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
netaddr==0.7.19
netifaces==0.10.4
networkx==2.5.1
notebook==6.4.10
numba==0.53.1
numpy==1.19.5
nvsm==20.3.6
nvsm-dshm==20.3.6
nvsm-health==20.3.6
nvsm-sdk==1.0.0
oauth==1.0.1
oauthlib==3.2.0
olefile==0.45.1
opencv-contrib-python==4.6.0.66
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.34
opt-einsum==3.3.0
p-tqdm==1.4.0
packaging==21.3
paho-mqtt==1.3.1
PAM==0.4.2
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathos==0.2.8
pathspec==0.9.0
pexpect==4.2.1
pickleshare==0.7.5
Pillow==8.4.0
pkg-about==1.0.3
pooch==1.6.0
pox==0.3.0
ppft==1.6.6.4
progressbar33==2.4
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.4.2
ptflops==0.6.9
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.2
pygame==2.1.2
Pygments==2.12.0
PyGObject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.9
pyRFC3339==1.0
pyshark==0.4.6
PySocks==1.7.1
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python-Levenshtein==0.12.2
pytorch-warmup==0.0.4
pytz==2022.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==23.1.0
qtconsole==5.2.2
QtPy==2.0.1
reportlab==3.4.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
resampy==0.2.2
rsa==4.8
scapy==2.4.5
scapy-ssl-tls==1.2.3
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
Shapely==1.7.0
simplejson==3.13.2
six==1.15.0
sklearn==0.0
sos==4.3
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tabulate==0.8.6
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.2
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.6.12
tinyec==0.4.0
torch==1.8.0
torch-scatter==2.0.9
torchvision==0.9.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wavio==0.0.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.44.0
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wrapt==1.12.1
xkit==0.0.0
xmljson==0.2.0
yapf==0.32.0
zipp==3.6.0
zope.interface==4.3.2
